form class

+++ includes/form.php
@@ -10,7 +10,8 @@ class Form {
 	public $field_types = array(
 		"alphaNumeric" => "text"
 		);
-	public $validation_errors;
+	
+	public $validation_errors = array();
 
 	function __construct($class_name, $fields=[]){
 		if(is_object($class_name)) {
@@ -40,6 +41,7 @@ class Form {
 		}
 		$this->render_button();
 		$this->render_form_end();
+
 	}
 
 	public function render_form_begin(){
@@ -64,31 +66,60 @@ class Form {
 		} else {
 			$label = "No label";
 		}
-		echo "<div class=\"form-group\">";
-		echo "<label for=\"{$field}\">{$label}</label>";
-		echo "<input type=\"text\" class=\"form-control\" name=\"{$field}\" value=\"{$value}\"/>";
-		echo "</div>";
+
+		// echo $this->validation_errors[$field];
+
+
+
+
+		if (array_key_exists ($field , $this->validation_errors)){
+			echo "<div class=\"form-group has-error has-feedback\">";
+			echo "<label for=\"{$field}\">{$label}</label>";
+			echo "<input type=\"text\" class=\"form-control\" name=\"{$field}\" value=\"{$value}\"/>";
+			echo "<span class=\"glyphicon glyphicon-remove form-control-feedback\"></span>";
+			echo "</div>";
+		} else {
+			echo "<div class=\"form-group has-success has-feedback\">";
+			echo "<label for=\"{$field}\">{$label}</label>";
+			echo "<input type=\"text\" class=\"form-control\" name=\"{$field}\" value=\"{$value}\"/>";
+			echo "</div>";
+		}
+
+
+
+
+
+
+
 	}
 
 	// populate class with POST elements
 	public function parsePost($post_array) {
 		foreach ($this->fields as $field) {
 			$this->model_class->$field = $post_array[$field];
+			# Validate field according to rules
+			# And populate validation_array if found errors
+			$this->validate_field($field, $this->model_class->$field);
 		}
 		return $this->model_class;
 	}
 
-	public function validate_fields() {
-		foreach ($this->fields as $field) {
-			echo "Validating field " . $field . "<br/>";
-			$this->validate_field( $field , $this->model_class->$field);
-		}
-		return true;
+
+	public function has_validation_errors() {
+		# Comment
+		return empty($this->validation_errors) ? false : true;
 	}
+
+	// public function validate_fields() {
+	// 	foreach ($this->fields as $field) {
+	// 		echo "Validating field " . $field . "<br/>";
+	// 		$this->validate_field( $field , $this->model_class->$field);
+	// 	}
+	// 	return true;
+	// }
 
 	public function validate_field( $field, $value) {
 		$validation_rules = $this->validations[$field];
-		print_r($validation_rules);
 
 		if( array_key_exists("allowEmpty", $validation_rules) ){
 			if ( strlen($value) == 0 and !$validation_rules["allowEmpty"] ){
@@ -111,10 +142,6 @@ class Form {
 				return;
 			}
 		}
-
-
-
-
 	}
